created Module Section Editor Page component for admin panel
