+++ components/common/pages/admin_panel/module_section_editor/module_section_editor.tsx
@@ -1,17 +1,29 @@
 import { useState } from 'react'
+import { ModuleLesson } from '../module_page_editor/lesson/lesson'
+import { ModuleSection } from '../module_page_editor/section/section'
 export const ModuleSectionEditor = () => {
 	const [isclicked, setIsClicked] = useState(false)
 	const showContent = () => setIsClicked(!isclicked)
 	return (
 		<>
-			<button
-				className="sectionToggle border-solid border-gray bg-gray-200  hover:bg-gray-100 rounded border text-2xl "
-				onClick={showContent}
+			<div
+				className="sectionToggle text-2xl "
 			>
-				<span className="m-2">Section Component</span>
-			</button>
-			<div className="sectionLessons">
-				{isclicked ? <div>Lesson Component</div> : ''}
+                    <span onClick={showContent} className='inline-block'>
+                    <ModuleSection
+                    sectionId={1}
+                    sectionName="Section 1"
+                    />
+                    </span>
+			</div>
+			<div className="sectionLessons ml-5">
+				{isclicked ? 
+                            <ModuleLesson
+                            lessonId={1}
+                            lessonName="Introduction to Dynamic Web Content"
+                            /> 
+                            : 
+                            ''}
 			</div>
 		</>
 	)
